+++ choco/src/main/java/btrplace/solver/choco/actionModel/ShutdownableNodeModel.java
@@ -24,9 +24,11 @@ import btrplace.solver.SolverException;
 import btrplace.solver.choco.NodeActionModel;
 import btrplace.solver.choco.ReconfigurationProblem;
 import btrplace.solver.choco.chocoUtil.FastIFFEq;
-import btrplace.solver.choco.chocoUtil.FastImpliesEq;
+import btrplace.solver.choco.chocoUtil.FastImpliesEq0;
 import choco.cp.solver.CPSolver;
 import choco.cp.solver.constraints.integer.ElementV;
+import choco.cp.solver.variables.integer.BoolVarNot;
+import choco.cp.solver.variables.integer.BooleanVarImpl;
 import choco.kernel.solver.variables.integer.IntDomainVar;
 
 import java.util.UUID;
@@ -112,10 +114,9 @@ public class ShutdownableNodeModel implements NodeActionModel {
             - If the node is hosting running VMs, it is necessarily online
             - If the node is offline, it is sure it cannot host any running VMs
         */
-        isOnline = s.createBooleanVar(rp.makeVarLabel(new StringBuilder("shutdownnableNode(").append(e).append(").online").toString()));
-        isOffline = s.createBooleanVar(rp.makeVarLabel(new StringBuilder("shutdownnableNode(").append(e).append(").offline").toString()));
-        s.post(s.neq(isOnline, isOffline));
-        s.post(new FastImpliesEq(isOffline, rp.getNbRunningVMs()[rp.getNode(e)], 0));
+        isOnline = s.createBooleanVar(rp.makeVarLabel(new StringBuilder("shutdownableNode(").append(e).append(").online").toString()));
+        isOffline = new BoolVarNot(s, new StringBuilder("shutdownableNode(").append(e).append(").offline").toString(), (BooleanVarImpl) isOnline);
+        s.post(new FastImpliesEq0(isOffline, rp.getNbRunningVMs()[rp.getNode(e)]));
 
         /*
         * D = {0, d}
@@ -140,7 +141,7 @@ public class ShutdownableNodeModel implements NodeActionModel {
         hostingStart = rp.getStart();
         //The moment the node can no longer host VMs varies depending on its next state
         hostingEnd = rp.makeDuration(new StringBuilder("shutdownableNode(").append(e).append(").hostingEnd").toString());
-        s.post(s.leq(hostingEnd, rp.getEnd()));
+        //s.post(s.leq(hostingEnd, rp.getEnd()));
 
         /*
           T = { As, RP.end}