+++ choco/src/main/java/btrplace/solver/choco/objective/minMTTR/RandomVMPlacement.java
@@ -139,7 +139,7 @@ public class RandomVMPlacement implements ValSelector<IntDomainVar> {
         if (stay) {
             UUID vm = vmPlacement.get(x);
             if (VMPlacementUtils.canStay(rp, vm)) {
-                rp.getLogger().debug("{} - VM {} stays on {} ({})", dbgLbl, vm, rp.getSourceModel().getMapping().getVMLocation(vm), rp.getNode(rp.getSourceModel().getMapping().getVMLocation(vm)));
+                //rp.getLogger().debug("{} - VM {} stays on {} ({})", dbgLbl, vm, rp.getSourceModel().getMapping().getVMLocation(vm), rp.getNode(rp.getSourceModel().getMapping().getVMLocation(vm)));
                 return rp.getNode(rp.getSourceModel().getMapping().getVMLocation(vm));
             }
         }
@@ -151,10 +151,10 @@ public class RandomVMPlacement implements ValSelector<IntDomainVar> {
             } else {
                 nIdx = randomValue(x);
             }
-            rp.getLogger().debug("{} - VM {} move to {} ({})", dbgLbl, vmPlacement.get(x), rp.getNode(nIdx), nIdx);
+            //rp.getLogger().debug("{} - VM {} move to {} ({})", dbgLbl, vmPlacement.get(x), rp.getNode(nIdx), nIdx);
             return nIdx;
         }
-        rp.getLogger().debug("{} - {} ", dbgLbl, x);
+        //rp.getLogger().debug("{} - {} ", dbgLbl, x);
         return x.getVal();
     }
 }